+++ providers/gitlab.go
@@ -1,16 +1,21 @@
 package providers
 
 import (
+	"fmt"
 	"log"
 	"net/http"
 	"net/url"
+	"strings"
 
+	"github.com/bitly/go-simplejson"
 	"github.com/pusher/oauth2_proxy/api"
 )
 
 // GitLabProvider represents an GitLab based Identity Provider
 type GitLabProvider struct {
 	*ProviderData
+	Group        string
+	EmailDomains []string
 }
 
 // NewGitLabProvider initiates a new GitLabProvider
@@ -34,28 +39,89 @@ func NewGitLabProvider(p *ProviderData) *GitLabProvider {
 		p.ValidateURL = &url.URL{
 			Scheme: "https",
 			Host:   "gitlab.com",
-			Path:   "/api/v4/user",
+			Path:   "/api/v4",
 		}
 	}
-	if p.Scope == "" {
-		p.Scope = "read_user"
-	}
+	p.Scope += " read_user"
 	return &GitLabProvider{ProviderData: p}
 }
 
 // GetEmailAddress returns the Account email address
 func (p *GitLabProvider) GetEmailAddress(s *SessionState) (string, error) {
+	if p.Group != "" {
+		groupsJSON, groupsJSONError := data(p.ValidateURL.String()+"/groups?access_token="+s.AccessToken, s)
+		if groupsJSONError != nil {
+			return "", groupsJSONError
+		}
+		groupsValid := false
+		for i := range groupsJSON.MustArray() {
+			group, groupJSONError := groupsJSON.GetIndex(i).Get("name").String()
+			if groupJSONError == nil {
+				groupsValid = groupsValid || (strings.ToLower(group) == p.Group)
+			}
+		}
+		if !groupsValid {
+			return "", fmt.Errorf("user has no acces to group %s", p.Group)
+		}
+	}
 
-	req, err := http.NewRequest("GET",
-		p.ValidateURL.String()+"?access_token="+s.AccessToken, nil)
+	singlemailJSON, singlemailJSONError := data(p.ValidateURL.String()+"/user?access_token="+s.AccessToken, s)
+	if singlemailJSONError != nil {
+		return "", singlemailJSONError
+	}
+	singlemail, singlemailError := singlemailJSON.Get("email").String()
+	if singlemailError != nil {
+		return "", singlemailError
+	}
+
+	if (len(p.EmailDomains) == 0) || (p.EmailDomains[0] == "*") {
+		return singlemail, nil
+	}
+
+	emailCandidates := []string{
+		singlemail,
+	}
+	emailsJSON, emailsError := data(p.ValidateURL.String()+"/user/emails?access_token="+s.AccessToken, s)
+	if emailsError != nil {
+		return "", emailsError
+	}
+	for i := range emailsJSON.MustArray() {
+		objmail, objmailError := emailsJSON.GetIndex(i).Get("email").String()
+		if objmailError == nil {
+			emailCandidates = append(emailCandidates, objmail)
+		}
+	}
+	for _, emailCandidate := range emailCandidates {
+		for _, domain := range p.EmailDomains {
+			if strings.HasSuffix(emailCandidate, domain) {
+				return emailCandidate, nil
+			}
+		}
+	}
+	return emailCandidates[0], nil
+}
+
+// SetGroup adds api scope to the oidc scopes
+func (p *GitLabProvider) SetGroup(group string) {
+	p.Group = strings.ToLower(group)
+	p.Scope += " api"
+}
+
+// SetEmailDomains to filter emails for
+func (p *GitLabProvider) SetEmailDomains(domains []string) {
+	p.EmailDomains = domains
+}
+
+func data(url string, s *SessionState) (*simplejson.Json, error) {
+	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		log.Printf("failed building request %s", err)
-		return "", err
+		return nil, err
 	}
 	json, err := api.Request(req)
 	if err != nil {
 		log.Printf("failed making request %s", err)
-		return "", err
+		return nil, err
 	}
-	return json.Get("email").String()
+	return json, nil
 }