+++ providers/gitlab_test.go
@@ -27,17 +27,30 @@ func testGitLabProvider(hostname string) *GitLabProvider {
 	return p
 }
 
-func testGitLabBackend(payload string) *httptest.Server {
-	path := "/api/v4/user"
+func testGitLabBackend(payloadUser string, payloadGroups string, payloadUserEmail string) *httptest.Server {
 	query := "access_token=imaginary_access_token"
+	userPath := "/api/v4/user"
+	groupsPath := "/api/v4/groups"
+	userEmailsPath := "/api/v4/user/emails"
 
 	return httptest.NewServer(http.HandlerFunc(
 		func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path != path || r.URL.RawQuery != query {
-				w.WriteHeader(404)
-			} else {
+			if r.URL.RawQuery != query {
+				w.WriteHeader(401)
+				return
+			}
+			switch r.URL.Path {
+			case userPath:
 				w.WriteHeader(200)
-				w.Write([]byte(payload))
+				w.Write([]byte(payloadUser))
+			case userEmailsPath:
+				w.WriteHeader(200)
+				w.Write([]byte(payloadUserEmail))
+			case groupsPath:
+				w.WriteHeader(200)
+				w.Write([]byte(payloadGroups))
+			default:
+				w.WriteHeader(404)
 			}
 		}))
 }
@@ -50,9 +63,9 @@ func TestGitLabProviderDefaults(t *testing.T) {
 		p.Data().LoginURL.String())
 	assert.Equal(t, "https://gitlab.com/oauth/token",
 		p.Data().RedeemURL.String())
-	assert.Equal(t, "https://gitlab.com/api/v4/user",
+	assert.Equal(t, "https://gitlab.com/api/v4",
 		p.Data().ValidateURL.String())
-	assert.Equal(t, "read_user", p.Data().Scope)
+	assert.Equal(t, " read_user", p.Data().Scope)
 }
 
 func TestGitLabProviderOverrides(t *testing.T) {
@@ -79,11 +92,11 @@ func TestGitLabProviderOverrides(t *testing.T) {
 		p.Data().RedeemURL.String())
 	assert.Equal(t, "https://example.com/api/v4/user",
 		p.Data().ValidateURL.String())
-	assert.Equal(t, "profile", p.Data().Scope)
+	assert.Equal(t, "profile read_user", p.Data().Scope)
 }
 
 func TestGitLabProviderGetEmailAddress(t *testing.T) {
-	b := testGitLabBackend("{\"email\": \"michael.bland@gsa.gov\"}")
+	b := testGitLabBackend("{\"email\": \"michael.bland@gsa.gov\"}", "", "")
 	defer b.Close()
 
 	bURL, _ := url.Parse(b.URL)
@@ -98,7 +111,7 @@ func TestGitLabProviderGetEmailAddress(t *testing.T) {
 // Note that trying to trigger the "failed building request" case is not
 // practical, since the only way it can fail is if the URL fails to parse.
 func TestGitLabProviderGetEmailAddressFailedRequest(t *testing.T) {
-	b := testGitLabBackend("unused payload")
+	b := testGitLabBackend("unused payload", "", "")
 	defer b.Close()
 
 	bURL, _ := url.Parse(b.URL)
@@ -114,11 +127,54 @@ func TestGitLabProviderGetEmailAddressFailedRequest(t *testing.T) {
 }
 
 func TestGitLabProviderGetEmailAddressEmailNotPresentInPayload(t *testing.T) {
-	b := testGitLabBackend("{\"foo\": \"bar\"}")
+	b := testGitLabBackend("{\"foo\": \"bar\"}", "", "")
 	defer b.Close()
 
 	bURL, _ := url.Parse(b.URL)
 	p := testGitLabProvider(bURL.Host)
+
+	session := &SessionState{AccessToken: "imaginary_access_token"}
+	email, err := p.GetEmailAddress(session)
+	assert.NotEqual(t, nil, err)
+	assert.Equal(t, "", email)
+}
+
+func TestGitLabProviderGetEmailAddressWithEmailDomain(t *testing.T) {
+	b := testGitLabBackend("{\"email\": \"ruben.wagner@example.com\"}", "", "[{\"email\": \"ruben.wagner@example2.com\"}]")
+	defer b.Close()
+
+	bURL, _ := url.Parse(b.URL)
+	p := testGitLabProvider(bURL.Host)
+	p.SetEmailDomains([]string{"@example2.com"})
+
+	session := &SessionState{AccessToken: "imaginary_access_token"}
+	email, err := p.GetEmailAddress(session)
+	assert.Equal(t, nil, err)
+	assert.Equal(t, "ruben.wagner@example2.com", email)
+}
+func TestGitLabProviderGetEmailAddressWithGroups(t *testing.T) {
+	b := testGitLabBackend("{\"email\": \"ruben.wagner@example.com\"}", "[{\"name\": \"testgroup\"}, {\"name\": \"testgroup2\"}]", "")
+	defer b.Close()
+
+	bURL, _ := url.Parse(b.URL)
+	p := testGitLabProvider(bURL.Host)
+	p.SetGroup("testgroup")
+	assert.Equal(t, " read_user api", p.Scope)
+
+	session := &SessionState{AccessToken: "imaginary_access_token"}
+	email, err := p.GetEmailAddress(session)
+	assert.Equal(t, nil, err)
+	assert.Equal(t, "ruben.wagner@example.com", email)
+}
+
+func TestGitLabProviderGetEmailAddressWithGroupsInvalid(t *testing.T) {
+	b := testGitLabBackend("{\"email\": \"ruben.wagner@example.com\"}", "[{\"name\": \"testgroup2\"}]", "")
+	defer b.Close()
+
+	bURL, _ := url.Parse(b.URL)
+	p := testGitLabProvider(bURL.Host)
+	p.SetGroup("testgroup")
+	assert.Equal(t, " read_user api", p.Scope)
 
 	session := &SessionState{AccessToken: "imaginary_access_token"}
 	email, err := p.GetEmailAddress(session)